[Refactor] request handler mapper의 getRequestHandler 기능을 추상화하여 interface 추출
